+++ src/LinearManhattanDisplay/TooltipComponent.tsx
@@ -1,41 +1,28 @@
 import React from 'react'
 
+import { getConf } from '@jbrowse/core/configuration'
+import { SanitizedHTML } from '@jbrowse/core/ui'
 import { Tooltip } from '@jbrowse/plugin-wiggle'
 import { observer } from 'mobx-react'
 
+import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
 import type { Feature } from '@jbrowse/core/util'
 import type { TooltipContentsComponent } from '@jbrowse/plugin-wiggle'
 
-const en = (n: number) => n.toLocaleString('en-US')
-
-function toP(s = 0) {
-  return +(+s).toPrecision(6)
-}
 interface Props {
   feature: Feature
+  model: Model
+}
+interface Model {
+  configuration: AnyConfigurationModel
+  featureUnderMouse?: Feature
 }
 
 const TooltipContents = React.forwardRef<HTMLDivElement, Props>(
-  function TooltipContents2({ feature }, ref) {
-    const start = feature.get('start') + 1
-    const end = feature.get('end')
-    const refName = feature.get('refName')
-    const name = feature.get('name')
-    const rsid = feature.get('rsid')
-    const loc = [
-      refName,
-      start === end ? en(start) : `${en(start)}..${en(end)}`,
-    ]
-      .filter(f => !!f)
-      .join(':')
-
+  function TooltipContents2({ model, feature }, ref) {
     return (
       <div ref={ref}>
-        {loc}
-        <br />
-        {`${toP(feature.get('score'))}`}
-        <br />
-        {name || rsid}
+        <SanitizedHTML html={getConf(model, 'mouseover', { feature })} />
       </div>
     )
   },
@@ -44,9 +31,7 @@ const TooltipContents = React.forwardRef<HTMLDivElement, Props>(
 type Coord = [number, number]
 
 const TooltipComponent = observer(function (props: {
-  model: {
-    featureUnderMouse?: Feature
-  }
+  model: Model
   height: number
   offsetMouseCoord: Coord
   clientMouseCoord: Coord
